product abd warehouse

+++ database/migrations/2023_05_09_111523_create_units_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('type')->nullable();
             $table->string('group');
-            $table->float('conversion_factor',16);
+            $table->double('conversion_factor');
             $table->boolean('active')->default(true);
             $table->timestamps();
         });
